legacy/requirements.txt
pandas
torch==1.6.0+cu101 -f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.7.0+cu101 -f https://download.pytorch.org/whl/torch_stable.html
scikit-learn
pytorch-lightning==0.10.0
torchaudio
